% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bs_seq_toy_matrix}
\alias{bs_seq_toy_matrix}
\title{BS-Seq Toy data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4510 rows and 20 columns.
}
\usage{
bs_seq_toy_matrix
}
\description{
Example of tumor and control beta values from Bisulphite Sequencing
}
\keyword{datasets}
