% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bs_toy_indexes}
\alias{bs_toy_indexes}
\title{Toy data}
\format{An object of class \code{list} of length 1000.}
\usage{
bs_toy_indexes
}
\description{
Example of data to map CpG sites to single CpG islands
}
\keyword{datasets}
