% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bs_toy_sites}
\alias{bs_toy_sites}
\title{BS Toy sites}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4510 rows and 2 columns.}
\usage{
bs_toy_sites
}
\description{
Example of data.frame with location of CpG sites
}
\keyword{datasets}
