% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_sites.R
\name{clusters_reduction}
\alias{clusters_reduction}
\title{Avoid retrieving of CpG sites too close to each others}
\usage{
clusters_reduction(sites_idx, N, min_dist)
}
\arguments{
\item{sites_idx}{a vector of integers}

\item{N}{number of sites to retrieve}

\item{min_dist}{an integer. Distance in basepairs}
}
\value{
a vector of indexes with close sites removed.
}
\description{
Remove sites within 'min_dist' (keep only one, per "cluster"), keeping at most N sites
}
\keyword{internal}
