% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AUC.R
\name{compute_AUC}
\alias{compute_AUC}
\title{Compute Area Under Curve for a matrix of samples}
\usage{
compute_AUC(tumor_table, control_table, ncores = 1, na_threshold = 0)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (percentage) from tumor samples.}

\item{control_table}{A matrix of beta-values (percentage) from
normal/control samples.}

\item{ncores}{Number of parallel processes to use for parallel computing.}

\item{na_threshold}{Fraction of NAs (considered independently in tumor and
control samples) above which a site will not be selected (default=0).}
}
\value{
A vector of AUC scores.
}
\description{
This function computes the Area Under Curve that is then used to define the
segregation between tumor and control samples accordingly to their
methylation (beta) values.
}
