% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_islands_indexes.R
\name{compute_island_indexes}
\alias{compute_island_indexes}
\title{Associated to each cpg islands the corresponding cpg site}
\usage{
compute_island_indexes(cpg_sites, cpg_islands_df = cpg_islands)
}
\arguments{
\item{cpg_sites}{a data.frame reporting the location of CpG sites from Bisulphite
Sequencing data and requiring two columns: chromosome and 1-based coordinate of CpG sites.}

\item{cpg_islands_df}{a data.frame reporting the location of CpG islands. By default it
loads the `cpg_islands` data.frame that comes with the package (based on data
downloaded from [UCSC Genome Browser](https://genome.ucsc.edu/) on date 2016-03-01.}
}
\description{
Associated to each cpg islands the corresponding cpg site
}
\examples{
\dontrun{
cpg_indexes <- compute_islands_indexes(bs_toy_sites)
}
\donttest{
cpg_indexes_updated <- compute_islands_indexes(bs_toy_sites, cpg_islands_df = cpg_islands_v2)
}
}
