% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_purity.R
\name{compute_purity}
\alias{compute_purity}
\title{Compute Purity}
\usage{
compute_purity(tumor, list_of_sites)
}
\arguments{
\item{tumor}{A matrix of beta scores generated from an Illumina
Methylation BeadChip (rows=sites; columns=samples)}

\item{list_of_sites}{A list of indexes generated by `select_informative_sites`
to be used to compute purity}
}
\value{
A vector of purity estimates.
}
\description{
Determine normal sample admixture from a set of tumor samples.
}
\examples{
compute_purity (tumor, list_of_sites)
}
