% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_toy_data}
\alias{control_toy_data}
\title{Toy data}
\format{An object of class \code{matrix} with 27578 rows and 38 columns.}
\usage{
control_toy_data
}
\description{
Example of control beta values from Illumina BeadChip
}
\keyword{datasets}
