% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpg_islands}
\alias{cpg_islands}
\title{Deprecated data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26662 rows and 4 columns.}
\usage{
cpg_islands
}
\description{
Dataset has been renamed, use \code{cpg_islands_df} instead.
Please change your code to use the new name, as
the deprecated dataset will become defunct in future versions.
}
\keyword{datasets}
