% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AUC.R
\name{get_AUC}
\alias{get_AUC}
\title{Calculate Area Under Curve comparing tumor and control samples}
\usage{
get_AUC(
  tumor_table,
  control_table,
  max_NA_data = 0,
  cores = 1,
  full_info = FALSE
)
}
\arguments{
\item{tumor_table}{A matrix of beta-values from tumor samples.}

\item{control_table}{A matrix of beta-values from normal/control samples.}

\item{max_NA_data}{At most this fraction of NAs (in tumor and control data
independently) is permitted to calculate AUC of a site (default=1, i.e. only
sites without NAs).}

\item{cores}{Number of cores for parallel computing.}

\item{full_info}{If TRUE (default=FALSE) returns also sites infos (fraction
of NAs in tumor and control tables) without skipping probes with missing values.}
}
\value{
A data.frame with AUC scores and probes names
}
\description{
This function calculates the Area Under Curve that define the segregation
between tumor and control samples accordingly to their methylation (beta)
values.
}
\examples{
auc_data <- get_AUC(tumor_toy_data, control_toy_data)
}
