% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R
\name{get_purity}
\alias{get_purity}
\title{Calculate purity or tumor content samples}
\usage{
get_purity(tumor_table, info_sites)
}
\arguments{
\item{tumor_table}{A matrix of beta-values}

\item{info_sites}{A data.frame reporting a set of informative sites
generated by \code{\link{find_informative_sites}} or
\code{\link{find_informative_regions}}.}
}
\value{
A data.frame with purity scores.
}
\description{
Estimate the proportion of cancer cells in the admixture of cells
forming the tumor microenvironment or the fraction of ctDNA in cfDNA samples.
}
\examples{
purity <- get_purity(tumor_toy_data,
 info_sites=data.frame(Probe=c("CpG_00001","CpG_00010","CpG_00020",
     "CpG_00015","CpG_00030","CpG_00045"), Site_type=rep(c("hyper","hypo"), each=3)))
}
