% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illumina27k_hg38}
\alias{illumina27k_hg38}
\title{Illumina 27k BeadChip mapped to Genome Reference hg38}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27578 rows and 10 columns.
}
\source{
\url{https://support.illumina.com/array/array_kits/infinium_humanmethylation27_beadchip_kit.html.html}
}
\usage{
illumina27k_hg38
}
\description{
Automatically loaded by \code{\link{select_informative_sites}}.
}
\keyword{datasets}
