% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illumina450k_hg19}
\alias{illumina450k_hg19}
\title{Illumina 27k BeadChip mapped to Genome Reference hg19}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 482421 rows and 4 columns.
}
\source{
\url{https://support.illumina.com/array/array_kits/infinium_humanmethylation450_beadchip_kit.html.html}
}
\usage{
illumina450k_hg19
}
\description{
Automatically loaded by \code{\link{select_informative_sites}}.
}
\keyword{datasets}
