% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_islands.R
\name{map_to_islands}
\alias{map_to_islands}
\title{Convert beta values from Bisulphite Sequencing to single CpG island beta values}
\usage{
map_to_islands(bs_data, sites_df, islands_df = cpg_islands, min_CpGs = 3)
}
\arguments{
\item{bs_data}{a matrix of beta values from Bisulphite Sequencing data.}

\item{sites_df}{a data.frame reporting the location of CpG sites from Bisulphite
Sequencing data. It requires two columns, (1) chromosome (format: `chrN`)
and (2) 1-based coordinate of CpG site.}

\item{islands_df}{a data.frame reporting the location of CpG islands. By default it
loads the `cpg_islands` data.frame that comes with the package (based on data
downloaded from [UCSC Genome Browser](https://genome.ucsc.edu/) in 2016-03-01.}

\item{min_CpGs}{an integer (default to 3). Minimum number of CpG sites within
a single CpG island
required to compute the mapped beta value (return NA if below that number).}
}
\value{
a matrix of beta values (one row per CpG island).
}
\description{
Remap beta values from different CpG sites to single beta values
associated to one CpG island where CpG sites are located.
}
\details{
Bisulphite Sequencing is used to retrieve level of DNA methylation at
base resolution. Different experiments though may retrieve
different CpG sites. This function makes a direct comparison possible.
}
\examples{
\donttest{
mapped <- map_to_islands(bs_control_toy_data, bs_toy_sites)
}
\dontrun{
mapped <- map_to_islands(bs_tumor_toy_data, bs_toy_data_sites,
               islands_df = cpg_islands_new, min_CpGs=5)
}
}
