% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_islands.R
\name{median_island}
\alias{median_island}
\title{Convert CpG sites to one CpG island}
\usage{
median_island(x, n)
}
\arguments{
\item{x}{A matrix.}

\item{n}{Minimum required number of sites per island (return NA otherwise)}
}
\description{
If the number of sites is sufficient take the median value else
return NA.
}
\keyword{internal}
