% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_regions.R
\name{median_of_region}
\alias{median_of_region}
\title{Transform CpG sites to one CpG region}
\usage{
median_of_region(x, n)
}
\arguments{
\item{x}{A subset matrix.}

\item{n}{Minimum required number of sites per region (return NA otherwise).}
}
\value{
A vector
}
\description{
If the number of sites is sufficient take the median value else return NA.
}
\keyword{internal}
