% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMES.R
\docType{package}
\name{PAMES}
\alias{PAMES}
\alias{PAMES-package}
\title{PAMES: Purity Assessment through clonal MEthylation Sites}
\description{
The PAMES package provides four functions:
select_info_sites, select_info_islands, reduce_bs_data, compute_purity.
}
\section{Foo functions}{

whatever
}

