% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_islands.R
\name{reduce_island}
\alias{reduce_island}
\title{Reduce a matrix subset}
\usage{
reduce_island(cpg_sites_matrix, n)
}
\arguments{
\item{cpg_sites_matrix}{a matrix}

\item{n}{minimum required number of sites per island (return NA otherwise)}
}
\description{
Takes into consideration the eventual presence of NAs
}
\keyword{internal}
