% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_regions.R
\name{reduce_to_regions}
\alias{reduce_to_regions}
\title{Reduce beta values from CpG sites to genomic regions}
\usage{
reduce_to_regions(
  beta_table,
  cpg_sites,
  cpg_regions,
  min_CpGs = 3,
  method = c("median", "mean")
)
}
\arguments{
\item{beta_table}{A matrix of beta-values.}

\item{cpg_sites}{A data.frame reporting the genomic location of CpG sites.}

\item{cpg_regions}{A data.frame reporting the genomic location of genomic regions.}

\item{min_CpGs}{An integer (default=3). Minimum number of CpG sites
within a single genomic region (return NA otherwise).}

\item{method}{Function to summarise a region: either `median` or `mean` (default=median).}
}
\value{
A matrix of beta values, each row corresponding to one region.
}
\description{
Reduce several beta values from different CpG sites to a single beta value
associated to one genomic region (such as CpG islands).
Different technologies retrieve DNA methylation levels of
different CpG sites. This function makes possible a direct comparison.
}
\examples{
reduced_data <- reduce_to_regions(bs_seq_toy_matrix, bs_seq_toy_sites, cpg_islands[1:10,])
}
