% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_informative_sites.R
\name{remove_close_sites}
\alias{remove_close_sites}
\title{Remove close sites}
\usage{
remove_close_sites(sites_df, max_sites, min_distance)
}
\arguments{
\item{sites_df}{A data.frame}

\item{min_distanc}{A number}
}
\description{
Remove close sites
}
\keyword{internal}
