% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_islands.R
\name{select_informative_islands}
\alias{select_informative_islands}
\title{Select informative CpG islands}
\usage{
select_informative_islands(tumor_table, auc, max_sites = 20,
  hyper_range = c(min = 40, max = 90), hypo_range = c(min = 10, max =
  60), tumor)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (percentage) from tumor samples.}

\item{auc}{A vector of AUC scores generated by \code{compute_AUC}.}

\item{max_sites}{Maximum number of islands to retrieve (half hyper-, half
hypo-methylated) (default = 20).}

\item{hyper_range}{A vector of length 2 with minimum lower and upper values
required to select hyper-methylated informative sites.}

\item{hypo_range}{A vector of length 2 with minimum lower and upper values
required to select hypo-methylated informative sites.}
}
\value{
A named list of indexes of informative islands ("hyper-" and
"hypo-methylated").
}
\description{
This function generates a list of informative CpG islands to be used to
estimate the purity of a set of tumor samples.
}
\details{
Informative islands are divided into \code{hyper} and \code{hypo} depending
on their level of methylation with respect to the average beta-score of
normal samples. Both sets will be used to compute purity.
}
\examples{
reduced_data <- reduce_to_islands(round(bs_toy_matrix*100), bs_toy_indexes)
reduced_data <- round(reduced_data*100)
auc_data <- compute_AUC(reduced_data[,1:10], reduced_data[,11:20])
info_sites <- select_informative_islands(reduced_data[,1:10], auc_data)
}
