% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_regions.R
\name{select_informative_regions}
\alias{select_informative_regions}
\title{Select informative regions}
\usage{
select_informative_regions(tumor_table, auc, max_sites = 20,
  hyper_range = c(min = 40, max = 90), hypo_range = c(min = 10, max =
  60), method = c("even", "top", "hyper", "hypo"), percentiles = c(0,
  100), return_info = FALSE)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (percentage) from tumor samples.}

\item{auc}{A vector of AUC scores generated by \code{compute_AUC}.}

\item{max_sites}{Maximum number of regions to retrieve (half hyper-, half
hypo-methylated) (default = 20).}

\item{hyper_range}{A vector of length 2 with minimum lower and upper values
required to select hyper-methylated informative sites.}

\item{hypo_range}{A vector of length 2 with minimum lower and upper values
required to select hypo-methylated informative sites.}

\item{method}{How to select sites: "even" (half hyper-, half hypo-methylated sites),
"top" (highest AUC irregardless of hyper or hypomethylation), "hyper" (hyper-methylated sites only),
"hypo" (hypo-methylated, sites only).}

\item{percentiles}{Vector of length 2: lower and upper percentiles to
select sites with beta values outside hypo- and hyper-ranges (default =
0,100; 0th and 100th percentiles, i.e. only min and max beta should be outside of ranges).}
}
\value{
A named list of indexes of informative regions ("hyper-" and "hypo-methylated").
}
\description{
This function generates a list of informative regions to be used to estimate
the purity of a set of tumor samples.
}
\details{
Informative regions are divided into \code{hyper} and \code{hypo} depending
on their level of methylation with respect to the average beta-score of
normal samples. Both sets will be used to compute purity.
}
\examples{
reduced_data <- reduce_to_regions(bs_toy_matrix, bs_toy_sites, cpg_islands[1:1000,])
auc_data <- compute_AUC(reduced_data[,1:10], reduced_data[,11:20])
info_regions <- select_informative_regions(reduced_data[,1:10], auc_data)
}
