% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_sites.R
\name{select_informative_sites}
\alias{select_informative_sites}
\title{Select informative CpG sites}
\usage{
select_informative_sites(tumor, auc, max_sites = 20, min_distance = 1e+06,
  hyper_range = c(min = 0.4, max = 0.9), hypo_range = c(min = 0.1, max =
  0.6), genome = c("hg19", "hg38"), platform = c("450k", "27k"))
}
\arguments{
\item{tumor}{a matrix of beta scores generated by an Illumina BeadChip.}

\item{auc}{a vector of AUC scores generated by \code{compute_AUC}.}

\item{max_sites}{maximum number of sites to retrieve (half "hyper-", half "hypo-methylated")
(default = 20).}

\item{min_distance}{measured in bps (base pairs), used to avoid selection
of CpG sites located within such distancce from one another (default = 1e6).}

\item{hyper_range}{a vector of length two with minimum lower and upper values required to
select hyper-methylated informative sites}

\item{hypo_range}{a vector of length two with minimum lower and upper values required to
select hypo-methylated informative sites}

\item{genome}{which genome version to use to map probes and
exclude probes located too close to each other (either \strong{hg19} or \strong{hg38}).}

\item{platform}{which Illumina platform was the experiment perfromed on
(either \strong{450k} or \strong{27k}).}
}
\value{
a named list of indexes of informative sites ("hyper-" and "hypo-methylated").
}
\description{
This function generates a list of informative CpG sites to be used to estimate
the purity of a set of tumor samples.
}
\details{
Informative sites are divided into \code{hyper} and \code{hypo} depending
on their level of
methylation with respect to the average beta-score of normal samples.
Both sets will be used to compute purity.
}
\examples{
\donttest{
info_sites <- select_informative_sites(tumor_toy_data,
                                       auc_data,
                                       platform="27k")
info_sites.hg38 <- select_informative_sites(tumor_toy_data,
                                            auc_data,
                                            platform="27k",
                                            genome="hg38")
}
}
