% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_sites_ext.R
\name{select_informative_sites_ext}
\alias{select_informative_sites_ext}
\title{Select informative CpG sites (extended)}
\usage{
select_informative_sites_ext(
  tumor_table,
  control_table,
  auc,
  ref_table,
  max_sites = 20,
  min_distance = 1000000,
  percentiles = c(0, 100),
  hyper_range = c(min = 40, max = 90),
  hypo_range = c(min = 10, max = 60),
  control_costraints = c(30, 70),
  method = c("even", "top", "hyper", "hypo"),
  return_info = FALSE
)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (percentage) from tumor samples.}

\item{control_table}{A matrix of beta-values (percentage) from normal/control samples.}

\item{auc}{A vector of AUC scores generated by \code{compute_AUC}.}

\item{ref_table}{A data.frame with two columns (chromosome, genomic_coordinates).}

\item{max_sites}{Maximum number of sites to retrieve (half hyper-, half
hypo-methylated) (default = 20).}

\item{min_distance}{Avoid selection of CpG sites located within this
distance from one another (default = 1e6 bps).}

\item{percentiles}{Vector of length 2: lower and upper percentiles to
select sites with beta values outside hypo- and hyper-ranges
(default = 0,100; 0th and 100th percentiles,
i.e. only min and max beta should be outside of ranges).}

\item{hyper_range}{A vector of length 2 with minimum lower and upper values
required to select hyper-methylated informative sites.}

\item{hypo_range}{A vector of length 2 with minimum lower and upper values
required to select hypo-methylated informative sites.}

\item{control_costraints}{To select a site, "first quartile"/"third quartile" of control data must be above/below these beta-values.}

\item{method}{How to select sites: "even" (half hyper-, half hypo-methylated sites),
"top" (highest AUC irregardless of hyper or hypomethylation), "hyper" (hyper-methylated sites only),
"hypo" (hypo-methylated, sites only).}

\item{return_info}{Return also a data.frame of all informative sites (for debug purpose).}
}
\value{
A named list of indexes of informative sites ("hyper-" and "hypo-methylated").
}
\description{
This function generates a list of informative CpG sites to be used to estimate
the purity of a set of tumor samples.
}
\details{
The EXTENDED version was added to introduce the new parameter, named
\code{control_costraints}, to force selection of sites where upper/lower
quartiles of control scores are below beta-values given by
\code{control_costraints}.  Informative regions are divided into
\code{hyper} and \code{hypo} depending on their level of methylation with
respect to the average beta-score of normal samples. Both sets will be used
to compute purity.
}
\examples{
## WARNING: The following code doesn't retrieve any informative site
## Its purpose is to show how to use the tool
auc_data <- compute_AUC(tumor_toy_data, control_toy_data)
info_sites <- select_informative_sites(tumor_toy_data, auc_data, illumina27k_hg19[3:4])
}
