% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AUC.R
\name{single_AUC}
\alias{single_AUC}
\title{Compute AUC a single vector}
\usage{
single_AUC(scores, states, na_threshold)
}
\arguments{
\item{scores}{integer vector (range 1-100)}

\item{states}{logical vector (class labels)}

\item{na_threshold}{fraction of scores (range 0-1)}
}
\description{
Use Wilcoxon method to compute AUC.
Return NA if NA samples are more than threshold
}
\keyword{internal}
