% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_regions.R
\name{summarise_region}
\alias{summarise_region}
\title{Reduce many CpG sites to one CpG region}
\usage{
summarise_region(x, n, method)
}
\arguments{
\item{x}{A subset matrix.}

\item{n}{Minimum required number of sites per region (return NA otherwise).}

\item{method}{Either `median` or `mean`.}
}
\value{
A vector
}
\description{
If the number of sites is sufficient, take the median/mean value else return NA.
}
\keyword{internal}
