% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_sites.R
\name{too_many_NAs}
\alias{too_many_NAs}
\title{Check whether vector has too many NA values or not.}
\usage{
too_many_NAs(x, threshold)
}
\arguments{
\item{x}{a vector}

\item{threshold}{maximum fraction of x allowed to be NA}
}
\value{
logical
}
\description{
If set to 1 returns FALSE only if all values are NAs. If set to 0, returns TRUE
with a single NA
}
\keyword{internal}
