% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumor_toy_data}
\alias{tumor_toy_data}
\title{Toy data}
\format{An object of class \code{matrix} with 27578 rows and 297 columns.}
\usage{
tumor_toy_data
}
\description{
Test with \code{\link{select_informative_sites}}.
}
\keyword{datasets}
