% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bs_control_toy_data}
\alias{bs_control_toy_data}
\title{Toy data}
\format{An object of class \code{matrix} with 4510 rows and 10 columns.}
\usage{
bs_control_toy_data
}
\description{
Example of control beta values from Bisulphite Sequencing
}
\keyword{datasets}
