% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bs_toy_sites}
\alias{bs_toy_sites}
\title{Toy data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10000 rows and 2 columns.}
\usage{
bs_toy_sites
}
\description{
Test with \code{\link{reduce_to_islands}}.
}
\keyword{datasets}
