% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AUC.R
\name{compute_AUC}
\alias{compute_AUC}
\title{Compute Area Under Curve}
\usage{
compute_AUC(tumor, control, max_NAs_frac = 0.5, n_threads = 1)
}
\arguments{
\item{tumor}{a matrix of beta scores generated by an Illumina BeadChip.}

\item{control}{a matrix of beta scores generated by an Illumina BeadChip.}

\item{n_threads}{number of cores to use}

\item{max_NAs_fraction}{fraction of NAs (considered independently in tumor and
control samples) above which a site will not be selected (default=0.5).}
}
\value{
a vector of AUC scores
}
\description{
This function compute the Area Under Curve used to define the segregation
between tumor and normal samples according to their beta-score.
}
