% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AUC.R
\name{compute_AUC}
\alias{compute_AUC}
\title{Compute Area Under Curve}
\usage{
compute_AUC(tumor, control, max_NAs_frac = 1)
}
\arguments{
\item{tumor}{a matrix of beta scores generated by an Illumina BeadChip.}

\item{control}{a matrix of beta scores generated by an Illumina BeadChip.}

\item{max_NAs_frac}{a value between 0 and 1. Sites with a fraction of NAs
major than or equal to it are excluded (default = 1; no NAs allowed).}
}
\value{
a vector of AUC scores
}
\description{
This function compute the Area Under Curve used to define the segregation
between tumor and normal samples according to their beta-score.
}
