% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_islands_indexes.R
\name{compute_islands_indexes}
\alias{compute_islands_indexes}
\title{Associate CpG islands to corresponding CpG sites}
\usage{
compute_islands_indexes(cpg_sites, cpg_islands = cpg_islands_df)
}
\arguments{
\item{cpg_sites}{A data.frame reporting the location of CpG sites from Bisulphite
Sequencing data. It requires two columns: chromosome and 1-based coordinate of CpG sites.}

\item{cpg_islands}{A data.frame reporting the location of the CpG islands. By default it
loads the `cpg_islands_df` data.frame that comes with the package (based on data
downloaded from [UCSC Genome Browser](https://genome.ucsc.edu/) on date 2016-03-01.}
}
\description{
Associate CpG islands to corresponding CpG sites
}
\examples{
cpg_indexes_short <- compute_islands_indexes(bs_toy_sites, head(cpg_islands_df, 10))
\dontrun{
cpg_indexes <- compute_islands_indexes(bs_toy_sites)
}
}
