% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_purity.R
\name{compute_purity}
\alias{compute_purity}
\title{Compute Purity of Tumor Samples}
\usage{
compute_purity(tumor, list_of_sites)
}
\arguments{
\item{tumor}{a matrix of beta scores generated from an Illumina
BeadChip (rows=sites; columns=samples).}

\item{list_of_sites}{a list of indexes generated by \code{\link{select_informative_sites}}
to be used to compute purity.}
}
\value{
a vector of purity estimates.
}
\description{
Estimate the proportion of cancer cells in the admixture of cells
constituting tumor microenvironment.
}
\examples{
purity <- compute_purity(tumor_toy_data, list(hyper=c(1, 10, 20), hypo=c(15,30,45)))
}
