% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpg_islands}
\alias{cpg_islands}
\title{Curated set of CpG islands}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26662 rows and 4 columns.
}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
cpg_islands
}
\description{
Downloaded on 2016-03-01.
}
\keyword{datasets}
