% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMES.R
\docType{package}
\name{PAMES}
\alias{PAMES}
\alias{PAMES-package}
\title{PAMES: Purity Assessment from clonal MEthylation Sites}
\description{
The PAMES package provides a set of functions to estimate
the level of purity of tumor samples.
}
\details{
The two main functions of PAMES are \code{\link{select_informative_sites}}
(to retreive sites of interest), and \code{\link{compute_purity}} (to compute
purity of tumor samples providing a list of sites generated by
\code{\link{select_informative_sites}}). Additionally, if working with
Bisulphite Sequencing data, users must first
reduce the data with \code{\link{reduce_to_islands}}, than select informative
CpG islands with \code{\link{select_informative_islands}}, and finally
estimate purity with \code{\link{compute_purity}}.
}
