% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_islands.R
\name{reduce_to_islands}
\alias{reduce_to_islands}
\title{Convert beta values from Bisulphite Sequencing to single CpG island beta values}
\usage{
reduce_to_islands(cpg_sites_matrix, cpg_indexes, min_CpGs = 3, ncores = 1)
}
\arguments{
\item{cpg_sites_matrix}{A matrix of beta values.}

\item{cpg_indexes}{A list of indexes generated by
\code{compute_islands_indexes}.}

\item{min_CpGs}{An integer (default to 3). Minimum number of CpG sites
within a single CpG island required to compute the reduced beta value
(return NA otherwise).}

\item{ncores}{Number of parallel processes to use for parallel computing}
}
\value{
A matrix of beta values (nrow == length(cpg_indexes)).
}
\description{
Reduce beta values from different CpG sites to single beta values
associated to one CpG island where CpG sites are located.
}
\details{
Bisulphite Sequencing is used to retrieve level of DNA methylation at
base resolution. Different experiments may retrieve
different CpG sites. This function makes a direct comparison possible.
}
\examples{
reduced_tumor <- reduce_to_islands(bs_toy_matrix[,1:10], bs_toy_indexes)
reduced_control <- reduce_to_islands(bs_toy_matrix[,11:20], bs_toy_indexes)
}
