% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_regions.R
\name{reduce_to_regions}
\alias{reduce_to_regions}
\title{Reduce beta values from CpG sites to genomic regions}
\usage{
reduce_to_regions(beta_table, cpg_sites, cpg_regions, min_CpGs = 3)
}
\arguments{
\item{beta_table}{A matrix of beta-values (percentage).}

\item{cpg_sites}{A data.frame reporting the genomic location of CpG sites.}

\item{cpg_regions}{A data.frame reporting the genomic location of genomic regions.}

\item{min_CpGs}{An integer (default to 3). Minimum number of CpG sites
within a single genomic region required to compute the reduced beta value
(return NA otherwise).}
}
\value{
A matrix of beta values (nrow == length(cpg_indexes)).
}
\description{
Reduce several beta values from different CpG sites to single beta values
associated to one genomic region (such as CpG islands) where CpG sites are located.
Different technologies retrieve DNA methylation levels of
different CpG sites. This function makes a direct comparison possible.
}
\examples{
reduced_data <- reduce_to_regions(bs_toy_matrix, bs_toy_sites, cpg_islands[1:10,])
}
