% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_islands.R
\name{select_informative_islands}
\alias{select_informative_islands}
\title{Select informative CpG islands}
\usage{
select_informative_islands(tumor, auc, hyper_range = c(min = 0.4, max = 0.9),
  hypo_range = c(min = 0.1, max = 0.6), max_sites = 20)
}
\arguments{
\item{tumor}{a matrix of beta scores generated by BS technology.
Applied independently to each island (row).}

\item{max_sites}{maximum number of islands to retrieve (half hyper-,
half hypo-methylated) (default=20).}
}
\value{
a list of indexes of informative islands (divided into two sets:
"hyper"- and "hypo"-methylated).
}
\description{
Generate a list of informative CpG islands to be used to compute purity of
a set of tumor samples.
}
\details{
Informative islands are divided into \code{hyper} and \code{hypo} depending on their level of
methylation with respect to the average beta value of normal samples.
Both sets will be used to compute purity.
}
\examples{
\dontrun{
info_sites <- select_informative_islands(mapped_bs_tumor_toy_data, auc)
}
}
