% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_islands.R
\name{select_informative_islands}
\alias{select_informative_islands}
\title{Select CpG islands required to compute sample purity.}
\usage{
select_informative_islands(tumor, control, max_NAs_fraction = 0.5,
  max_sites = 20)
}
\arguments{
\item{tumor}{A matrix of beta scores generated by an Illumina MethylChip}

\item{control}{A matrix of beta scores generated by an Illumina MethylChip}

\item{max_NAs_fraction}{Fraction of NAs above which AUC will not be computed
(set to 1 to avoid sites with any NA). Applied independently to each sites (row).}

\item{max_sites}{Maximum number of sites to retrieve
(divided in two sets: "hyper-" and "hypo-methylated")}

\item{min_distance}{Measured in bp (base pair), used to avoid selection
of CpG sites located within this distance from another.}

\item{platform}{'27k' or '450k' Illumina HumanMethylation BeadChip.}

\item{genome}{Currently support hg19 or hg38. Used to map probes to genome and
remove heavly clusterd probes.}
}
\value{
A list indexes of informative sites (divided into two sets:
"hyper-methylated" and "hypo-methylated")
}
\description{
Generate a list of islands to be used to compute purity of a set of tumor samples.
Informative islands are divided in "hyper" and "hypo" depending on their level of
methylation with respect to the average of normal samples. Both sets will be used.
}
\examples{
select_informative_sites(tumor_matrix, control_matrix)
select_informative_sites(tumor_matrix, control_matrix, max_sites=40, min_distance=5000)
}
