% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_sites.R
\name{select_informative_sites}
\alias{select_informative_sites}
\title{Select informative CpG sites}
\usage{
select_informative_sites(tumor_table, auc, max_sites = 20,
  min_distance = 1000000, hyper_range = c(min = 40, max = 90),
  hypo_range = c(min = 10, max = 60), genome = c("hg19", "hg38"),
  platform = c("450k", "27k"), tumor)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (percentage) from tumor samples.}

\item{auc}{A vector of AUC scores generated by \code{compute_AUC}.}

\item{max_sites}{Maximum number of sites to retrieve (half hyper-, half
hypo-methylated) (default = 20).}

\item{min_distance}{To avoid selection of CpG sites located within such
distance from one another (default = 1e6 bps).}

\item{hyper_range}{A vector of length 2 with minimum lower and upper values
required to select hyper-methylated informative sites.}

\item{hypo_range}{A vector of length 2 with minimum lower and upper values
required to select hypo-methylated informative sites.}

\item{genome}{Genome version: used to exclude probes located too close to
each other (\strong{hg19} or \strong{hg38}).}

\item{platform}{Illumina platform used in the experiment (\strong{450k} or
\strong{27k}).}
}
\value{
A named list of indexes of informative sites ("hyper-" and "hypo-methylated").
}
\description{
This function generates a list of informative CpG sites to be used to estimate
the purity of a set of tumor samples.
}
\details{
Informative sites are divided into \code{hyper} and \code{hypo} depending on
their level of methylation with respect to the average beta-score of normal
samples. Both sets will be used to compute purity.
}
\examples{
auc_data <- compute_AUC(tumor_toy_data, control_toy_data)
## WARNING: The following code doesn't retrieve any informative site
## Its only purpose is to show how to use the tool
info_sites <- select_informative_sites(tumor_toy_data, auc_data, platform="27k")
info_sites.hg38 <- select_informative_sites(tumor_toy_data, auc_data,
 platform="27k", genome="hg38")
}
