% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_sites.R
\name{select_informative_sites}
\alias{select_informative_sites}
\title{Select CpG sites required to compute sample purity.}
\usage{
select_informative_sites(tumor, control, max_NAs_fraction = 0.5,
  max_sites = 20, min_distance = 1e+06, genome = c("hg19", "hg38"),
  platform = c("450k", "27k"))
}
\arguments{
\item{tumor}{A matrix of beta scores generated by an Illumina MethylChip}

\item{control}{A matrix of beta scores generated by an Illumina MethylChip}

\item{max_NAs_fraction}{Fraction of NAs above which AUC will not be computed
(set to 1 to avoid sites with any NA). Applied independently to each sites (row).}

\item{max_sites}{Maximum number of sites to retrieve
(divided in two sets: "hyper-" and "hypo-methylated")}

\item{min_distance}{Measured in bp (base pair), used to avoid selection
of CpG sites located within this distance from another.}

\item{genome}{Currently support hg19 or hg38. Used to map probes to genome and
remove heavly clusterd probes.}

\item{platform}{'27k' or '450k' Illumina HumanMethylation BeadChip.}
}
\value{
A list indexes of informative sites (divided into two sets:
"hyper-methylated" and "hypo-methylated")
}
\description{
Generate a list of sites to be used to compute purity of a set of tumor samples.
Informative sites are divided in "hyper" and "hypo" depending on their level of
methylation with respect to the average of normal sampels. Both sets will be used.
}
\examples{
select_informative_sites(tumor_matrix, control_matrix)
select_informative_sites(tumor_matrix, control_matrix, max_sites=40, min_distance=5000)
}
