% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_sites.R
\name{select_informative_sites}
\alias{select_informative_sites}
\title{Select informative CpG sites}
\usage{
select_informative_sites(tumor, control, max_NAs_fraction = 0.5,
  max_sites = 20, min_distance = 1e+06, genome = c("hg19", "hg38"),
  platform = c("450k", "27k"))
}
\arguments{
\item{tumor}{a matrix of beta scores generated by an Illumina BeadChip.}

\item{control}{a matrix of beta scores generated by an Illumina BeadChip.}

\item{max_NAs_fraction}{fraction of NAs above which a site will not be selected.
Applied independently to each site (row).}

\item{max_sites}{maximum number of sites to retrieve (half "hyper-", half "hypo-methylated")}

\item{min_distance}{measured in bp (base pair), used to avoid selection
of CpG sites located within this distance from one another.}

\item{genome}{\strong{hg19} or \strong{hg38}. Used to map probes to genome and
remove heavily clustered probes.}

\item{platform}{\strong{27k} or \strong{450k} Illumina BeadChip.}
}
\value{
a list of indexes of informative sites (divided into two sets:
"hyper-" and "hypo-methylated")
}
\description{
Generate a list of informative CpG sites to be used to compute purity of
a set of tumor samples.
}
\details{
Informative sites are divided into \code{hyper} and \code{hypo} depending
on their level of
methylation with respect to the average beta value of normal samples.
Both sets will be used to compute purity.
}
\examples{
\donttest{
info_sites <- select_informative_sites(tumor_toy_data,
                                       control_toy_data,
                                       platform="27k")
info_sites.hg38 <- select_informative_sites(tumor_toy_data,
                                            control_toy_data,
                                            platform="27k",
                                            genome="hg38")
}
}
