% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AUC.R
\name{single_AUC}
\alias{single_AUC}
\title{Compute AUC a single vector}
\usage{
single_AUC(x, state, na_threshold)
}
\arguments{
\item{x}{integer vector (range 0-1)}

\item{state}{logical vector}

\item{na_threshold}{numeric}
}
\description{
Return NA if NA samples are more than threshold
}
\keyword{internal}
