% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AUC.R
\name{single_AUC}
\alias{single_AUC}
\title{Calculate AUC}
\usage{
single_AUC(scores, is_tumor)
}
\arguments{
\item{scores}{integer vector (range 1-100)}

\item{is_tumor}{logical vector (class labels)}
}
\description{
Use Wilcoxon method to calculate AUC.
}
\keyword{\href{http://blog.revolutionanalytics.com/2017/03/auc-meets-u-stat.html}{http://blog.revolutionanalytics.com}}
\keyword{internal}
