% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_informative_sites.R
\name{too_close}
\alias{too_close}
\title{Check whether a site is too close to the other sites or not}
\usage{
too_close(new_idx, prev_idxs, min_distance, platform_data)
}
\arguments{
\item{new_idx}{an index (integer)}

\item{prev_idxs}{a vector of indexes (integer)}

\item{min_distance}{an integer. Distance in basepairs}
}
\value{
logical
}
\description{
Used in "cluster_recution" function only, given the index of a CpG site and
a set of indexes of other sites, check if last added site is less than
"min_distance" basepair distant from previously retrieved sites.
}
\keyword{internal}
