% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumor_toy_data}
\alias{tumor_toy_data}
\title{Illumina Tumor Toy data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 27578 rows and 297 columns.
}
\usage{
tumor_toy_data
}
\description{
Example of tumor beta values from Illumina BeadChip
}
\keyword{datasets}
